Reorganize toolbar
The stop button falled out of the toolbar, and we can't add more
functionality with the actual design. Now instead of use one button
for every type of globe, add one menuitem "Add globe" with options
by type. The icons were changed to remove the "+" sign because
is redundant. Separator removed from the toolbar too.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>