Copy textview window instead of use pango cairo text toy api - SL #4607
Cairo toy text api have problems, and can't render complex sripts.
Changed the way we draw the text in not edition mode to solve this
issue. This patch also move the textview, from be in the Box, to the
globe, then now we have one textview for ewvery globe instead of one by
box. This simplify tghe code.

Signed-off-by: Gonzalo Odiard <gonzalo@laptop.org>