Fix title box vertical position
The title box was getting occluded behind the first comic
box and image, due a wrong height calculation.

We can't trust that the widgets  height won't change,  after we
read its allocated height only once. But what we can do, is use
the fact that we are already giving that widget a fixed height.