Fix text rendering
In GTK3, the textview will only render what is visible
on screen.

For example, if the textview is contained in a scrolled
window, textview.get_window will return a drawable that
only shows the visible part of the textview.

The most common symptom is getting black rectangles.

One solution is to:
1. Let the textview draw directly to the cached
   text surface, and not through its window.
2. Avoid race conditions between imprimir and
   finalize_text methods, by setting textview
   initial size.
3. Explicitly draw the textview to the comic box
   context.